refactor: availabilitySettings atoms web wrapper and translations
